handled empty folders for baseline

+++ models/tf/baseline/GAN/out/clean_files.py
@@ -3,20 +3,30 @@ import glob
 
 files = glob.glob('./metric/*')
 for f in files:
+    if f[-5:] is 'empty':
+        continue
     os.remove(f)
 
 files = glob.glob('./evolution_hist/*')
 for f in files:
+    if f[-5:] is 'empty':
+        continue
     os.remove(f)
 
 files = glob.glob('./losses/*')
 for f in files:
+    if f[-5:] is 'empty':
+        continue
     os.remove(f)
 
 files = glob.glob('./saved_weights/*')
 for f in files:
+    if f[-5:] is 'empty':
+        continue
     os.remove(f)
 
 files = glob.glob('./tensorboard_files/*')
 for f in files:
+    if f[-5:] is 'empty':
+        continue
     os.remove(f)
